% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{compute_bayespin}
\alias{compute_bayespin}
\title{Calculates estimates of the probability of informed trading}
\usage{
compute_bayespin(mcmcest)
}
\arguments{
\item{mcmcest}{An \code{mcmcest} object of the \code{finmix} package containing
all estimated parameters from the finite mixture distribution of the
compressed EKOP model.}
}
\value{
A \code{data.frame} with estimated PINs from the maximum a posterior, the
Bayesian maximum likelihood and the identified ergodic average parameter
estimates of the underlying finite mixture distribution of the compressed
EKOP model.
}
\description{
Calling \code{\link[=compute_bayespin]{compute_bayespin()}} calculates the probability of informed trading
(PIN) from the paper of Easley et al. (1996). The input argument is an
\code{mcmcest} (see \code{\link[finmix]{mcmcest-class}}) object from the \code{finmix}
package containing all parameters estimates from the finite mixture
distribution of the compressed EKOP model in Grammig, Theissen and Zehnder
(2015).
}
\references{
\itemize{
\item Easley, D., Kiefer, N., O’Hara, M., Paperman, J., 1996. Liquidity,
information, and infrequently traded stocks. Journal of Finance 51,
1405–1436.
\item Grammig, J., Theissen, E., Zehnder, L.S., 2015. Bayesian Estimation of the
Probability of Informed Trading. Conference on Financial Econometrics &
Empirical Asset Pricing 2016, Lancaster University
}
}
\seealso{
\itemize{
\item \code{\link[finmix]{mcmcest-class}} for the definition of the \code{mcmcest} class union
\item \code{\link[=estimate_pin]{estimate_pin()}} for estimating the PIN with the Bayesian approach
described in Grammig, Theissen and Zehnder (2015)
\item \code{\link[=compute_pin]{compute_pin()}} for calculating the PIN for provided parameters
}
}
