% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_comp_lik}
\alias{compute_comp_lik}
\title{Computes the likelihood of the compressed EKOP model}
\usage{
compute_comp_lik(data, par, T, methodLik = c("precise", "approx"))
}
\arguments{
\item{data}{A \code{vector} containing the trades data.}

\item{par}{A vector specifying the parameter values at which the function
should be evaluated. The parameter order is \code{alpha}, \code{epsilon}, \code{delta},
and \code{mu}.}

\item{T}{A double indicating the length of a trading day in minutes.}

\item{methodLik}{A character specifying, if undefined function values in
optimization should be approximated by large defined values (\code{1e+6}). This
can help to make maximum likelihood estimation more stable.}
}
\value{
A double holding the likelihood value of the data and parameter
values passed in.
}
\description{
Calling \code{\link[=compute_comp_lik]{compute_comp_lik()}} computes the likelihood of the compressed EKOP
model. The compressed EKOP model has been first proposed by Jackson (2007)
and uses only the number of trades per trading day instead of the number of
buyer- and seller-initiated trades per day.
}
\details{
In case of large trading volumes the function computation suffers sometimes
from number overflow. In this case the parameter \code{methodLik} offers an
approximation method in which \code{Inf} or \code{NaN} values are replaced by the
value \code{1e+6} to avoid number overflow. In experiments this approximation has
shown very good performance and we suggest the user to consider this
approximation when function calculation fails to number overflow.
}
\references{
\itemize{
\item Jackson, D., 2007. Infering trader behavior from transaction data: A trade
count model. Journal of Computational and Graphical Statistics 12, 55-79.
}
}
\seealso{
\itemize{
\item \code{\link[=estimate_compml]{estimate_compml()}} for the calling function
}
}
