% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_ekop_lik}
\alias{compute_ekop_lik}
\title{Computes the modified likelihood from the EKOP model}
\usage{
compute_ekop_lik(data, par, T, methodLik = c("precise", "approx"))
}
\arguments{
\item{data}{A \code{data.frame} containing the trades data in the following order:
mis-specified buys, mis-specified sells, buyer-initiated trades,
seller-initiated trades and finally the total number of trades on a
specific trading day.}

\item{par}{A vector specifying the parameter values at which the function
should be evaluated. The parameter order is \code{alpha}, \code{epsilon}, \code{delta},
and \code{mu}.}

\item{T}{A double indicating the length of a trading day in minutes.}

\item{methodLik}{A character specifying, if undefined function values in
optimization should be approximated by large defined values (\code{1e+6}). This
can help to make maximum likelihood estimation more stable.}
}
\value{
A double holding the likelihood value of the data and parameter
values passed in.
}
\description{
Calling \code{\link[=compute_ekop_lik]{compute_ekop_lik()}} computes the modified likelihood function from
the model of Easley et al. (1996). The modifications are presented in the
paper by Easley et al. (2002) and are applied to the scenario of equal
buy and sell rates in the model.

Optimization is performed over the logit transformation of the ratios \code{alpha}
and \code{delta} and therefore these parameter logits get transformed in the
likelihood function via the logistic transformation \code{exp()/(1+exp())}.
}
\references{
\itemize{
\item Easley, D., Kiefer, N., O’Hara, M., Paperman, J., 1996. Liquidity,
information, and infrequently traded stocks. Journal of Finance 51,
1405–1436.
\item Easley, David, Hvidkjaer, Soeren, and O’Hara, Maureen (2002).
“Is Information Risk a Determinant of Asset Returns?” In: The Journal of
Finance 57.5, pp. 2185–2221. DOI: 10.1111/1540-6261.00493.
}
}
\seealso{
\itemize{
\item \code{\link[=estimate_mlekop]{estimate_mlekop()}} for the calling function
\item \code{\link[=compute_ekop_orig_lik]{compute_ekop_orig_lik()}} for the original likelihood function of the
paper by Easley et al. (1996)
}
}
