% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{compute_mlpin}
\alias{compute_mlpin}
\title{Computes the probability of informed trading from ML estimates}
\usage{
compute_mlpin(par)
}
\arguments{
\item{par}{A vector containing the parameter estimates from the optimization
procedure.}
}
\value{
A double holding the PIN estimate.
}
\description{
Calling \code{\link[=compute_mlpin]{compute_mlpin()}} returns the probability of informed trading as
presented in the paper of Easley et al. (1996).

The estimation procedure in \code{\link[=estimate_mlekop]{estimate_mlekop()}} uses the logits of the
rates \code{alpha} and \code{delta} to ensure a more stable optimization. To arrive
at the original parameter estimates the logistic transformation from
\code{\link[=logistic]{logistic()}} is applied.
}
\examples{
# Simulate trades data.
trades_data <- simulate_ekop()
# Estimate the EKOP model.
pin_optml <- estimate_mlekop(trades_data, methodLik="approx", 
                             fnLik="compute_ekop_orig_lik")
# Estimate the PIN from the parameter estimates.
compute_mlpin(pin_optml$par)

}
\references{
\itemize{
\item Easley, D., Kiefer, N., O’Hara, M., Paperman, J., 1996. Liquidity,
information, and infrequently traded stocks. Journal of Finance 51,
1405–1436.
}
}
\seealso{
\itemize{
\item \code{\link[=estimate_mlekop]{estimate_mlekop()}} for the calling function.
}
}
