% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{compute_pin}
\alias{compute_pin}
\title{Computes the probability of informed trading}
\usage{
compute_pin(alpha, epsilon, mu)
}
\arguments{
\item{alpha}{A double specifying the probability of an information event.}

\item{epsilon}{A double specifying the arrival rate of uninformed traders.}

\item{mu}{A double specifying the arrival rate of informed traders.}
}
\value{
The probability of informed trading.
}
\description{
Calling \code{\link[=compute_pin]{compute_pin()}} computes the probability of informed trading (PIN)
from the EKOP model for provided parameters. See Easley et al. (1996) for
details.
}
\examples{
compute_pin(.2, .25, .01)

}
\references{
\itemize{
\item Easley, D., Kiefer, N., O’Hara, M., Paperman, J., 1996. Liquidity,
information, and infrequently traded stocks. Journal of Finance 51,
1405–1436.
}
}
\seealso{
\itemize{
\item \code{\link[=compute_bayespin]{compute_bayespin()}} for computing the PIN from bayesian parameter
estimates
}
}
