% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{correct_trades}
\alias{correct_trades}
\title{Corrects mis-specified trades data}
\usage{
correct_trades(trades)
}
\arguments{
\item{trades}{A \code{data.frame} containing the trades data. This must have the
columns \code{MisBuy}, \code{MisSell}, \code{Buy}, and \code{Sell} specified.}
}
\value{
A \code{data.frame} containing the re-constructed well-specified trades
data.
}
\description{
Calling \code{\link[=correct_trades]{correct_trades()}} on simulated mis-specified trades data corrects
the mis-specifications. This becomes handy when using simulated data in
comparison studies of e.g. estimators, etc.
}
\details{
Mis-specified trades are recorded in the columns \code{MisBuy} and \code{MisSell} for
buy and sell trades respectively. This information is used to re-construct
the well-specified trades data.
}
\examples{
# Simulate mis-specifed trades data for 1000 trading dates with a trading 
# day lasting 6.5 hours.
trades_data <- simulate_ekop_mis(size = 1000, alpha = .3, epsilon = .4,
                                 delta = .5, mu = .04, T = 60*6.5, mis = .1)
trades_data_corr <- correct_trades(trades = trades_data)                        
head(trades_data_corr)

}
\seealso{
\itemize{
\item \code{\link[=simulateEKOPMis]{simulateEKOPMis()}} for simulating trades data with mis-specification
\item \code{\link[=simulateEKOP]{simulateEKOP()}} for simulating trades data without mis-specification
}
}
