% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_compml}
\alias{estimate_compml}
\title{Estimated the probability of informed trading of the compressed EKOP model}
\usage{
estimate_compml(
  data,
  startpar,
  T = 390,
  methodLik = c("precise", "approx"),
  fnscale = -1,
  trace = 0,
  grad_free = TRUE,
  return_opt = FALSE,
  opt_out = TRUE
)
}
\arguments{
\item{data}{A \code{vector} containing the number of trades per day.}

\item{startpar}{A vector containing start parameters for maximum likelihood
estimation. These must be starting values for the logit of alpha, epsilon,
and mu. If no starting values are provided the function chooses the values
\verb{(0, mean(trades)x.75/2, mean(trades)x.25/2)}.}

\item{T}{A double specifying the minutes of a trading day.}

\item{methodLik}{A character specifying, if undefined function values in
optimization should be approximated by large defined values (\code{1e+6}).
This can help to make maximum likelihood estimation more stable.}

\item{fnscale}{A negatve double specifying a scaling factor for the
likelihood function. This can in some cases help when the algorithm does
not reach convergence or suffer from number overflow.}

\item{trace}{An integer specifying which level of tracing should be used.
see \code{?optim} for more details.}

\item{grad_free}{A logical indicating if gradient-free optimization should
be used when gradient descent did not converge. If \code{TRUE} the optimization
procedure \code{\link[dfoptim]{nmkb}} is used.}

\item{return_opt}{A logical indicating, if in addition to the PIN estimates
also the results from the optimization procedure should be returned. If
\code{TRUE} a list is returned with an element \code{pin_estimates} holding the
PIN estimates and an element \code{opt_results} holding the output of the
optimization procedure.}

\item{opt_out}{(Deprecated) A logical indicating if only the output of the
optimization procedure should be returned. Some older applications still
rely on this output. In the next version this feature will be removed.
Note that the default value is \code{TRUE}.}
}
\value{
A \code{list} with all components as returned by
\code{\link[stats]{optim}} or \code{\link[dfoptim]{nmkb}}.
}
\description{
Calling \code{\link[=estimate_compml]{estimate_compml()}} estimates the probability of informed trading
(PIN) from the model of Jackson (2007) using maximum likelihood (ML). As
described in the paper by Jackson (2007) and similar to the Bayesian
estimation used in Grammig et al. (2015) this model needs only the number
of trades per day provided in the parameter \code{data}.
}
\details{
\subsection{Optimization algorithms}{

Optimization is performed gradient-based by using the
\code{\link[stats]{optim}} function. The algorithm used is the \code{L-BFGS-B}
that allows for parameter restrictions. This is necessary because we use for
the probability \code{alpha} the logistic transformation \code{exp()/(1+exp())} to
produce valid probabilities in optimization.

In case that the gradient-based algorithm does not converge in between 100
steps, a gradient-free optimization is applied. For gradient-free
optimization \code{\link[dfoptim]{nmkb}}, a bounded \code{Nelder-Mead} algorithm
is used. Derivative-free optimization is only performed after gradient-based
optimization did not converge and the argument \code{grad_free} is \code{TRUE}.

The argument \code{fnscale} can be used to scale the likelihood function in case
of very large values (e.g. very large volumes) that might lead to number
overflow during computation. Note, \code{fnscale} must always be negative as the
likelihood function should be maximized.
}
}
\examples{
# Simulate data from the EKOP model. 
trades_data <- simulateEKOP()
# Estimate the EKOP model by maximum likelihood.
pin_estml <- estimate_compml(trades_data$Trades, methodLik="approx", 
                             opt_out=FALSE)
                   
}
\references{
\itemize{
\item Jackson, D., 2007. Infering trader behavior from transaction data: A trade
count model. Journal of Computational and Graphical Statistics 12, 55-79.
\item Grammig, J., Theissen, E., Zehnder, L.S., 2015. Bayesian Estimation of the
Probability of Informed Trading. Conference on Financial Econometrics &
Empirical Asset Pricing 2016, Lancaster University
}
}
\seealso{
\itemize{
\item \code{\link[=estimate_pin]{estimate_pin()}} for estimating the PIN with a Bayesian approach that
needs only the total number of trades
\item \code{\link[=estimate_mlekop]{estimate_mlekop()}} for estimating the PIN with the original EKOP model
that needs the numbers of buyer- and seller-initiated trades per day
\item \code{\link[=compute_comp_lik]{compute_comp_lik()}} for the implementation of the likelihood function of
the paper of Jackson (2007).
}
}
