% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{logistic}
\alias{logistic}
\title{Transforms values by the logistic function}
\usage{
logistic(value)
}
\arguments{
\item{value}{A double or a vector of doubles to be transformed.}
}
\value{
A double or vector of doubles containing the logistic
transformations of the passed in arguments.
}
\description{
Calling \code{\link[=logistic]{logistic()}} applies the logistic transformation \code{exp()/(1+exp())}
to the passed in arguments.
}
\examples{
logistic(-1.5)

}
\seealso{
\itemize{
\item \code{\link[=computeEKOPLik]{computeEKOPLik()}} for a function applying the transformation during
optimization of parameters.
}
}
