% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateEKOPMis}
\alias{simulateEKOPMis}
\title{Simulates mis-specified trades data}
\usage{
simulateEKOPMis(
  size = 1000,
  alpha = 0.2,
  epsilon = 0.2,
  delta = 0.5,
  mu = 0.02,
  T = 60 * 6.5,
  mis = 0.15
)
}
\arguments{
\item{size}{An integer specifying the number of trading days that should
be simulated.}

\item{alpha}{A double specifying the probability of an information event.}

\item{epsilon}{A double specifying the arrival rate of uninformed traders.}

\item{delta}{A double specifying the probability of information being good
or bad.}

\item{mu}{A double specifying the arrival rate of informed traders.}

\item{T}{A double defining the length of a trading day in minutes.}

\item{mis}{A double specifying the mis-sepcification rate for trades.}
}
\value{
A \code{data.frame} containing the trades per day.
}
\description{
Simulates mis-specified trades data from the model by Easley, Kiefer, O'Hara
and Paperman (1996). The simulation function is implemented in C++ and allows
therefore usage in specific performance-sensitive scenarios, like simulating
huge amounts of trade data or calling this function repeatedly in large
simulations.
}
\details{
The functions returns a \code{data.frame} with named columns \code{MisBuy}, \code{MisSell},
\code{Buy}, \code{Sell}, and \code{Trades}. \code{Buy} and \code{Sell} are the number of buy and sell
trades per trading day and \code{Trades} defines the sum of these amounts.
\code{MisBuy} and \code{MisSell} define the number of mis-specified buy and sell
trades per trading day respectively.
}
\examples{
# Simulate mis-specifed trades data for 1000 trading dates with a trading 
# day lasting 6.5 hours.
trades_data <- simulateEKOPMis(size = 1000, alpha = .3, epsilon = .4,
                               delta = .5, mu = .04, T = 60*6.5, mis = .1)
head(trades_data)

}
\references{
\itemize{
\item Easley, D., Kiefer, N., O’Hara, M., Paperman, J., 1996. Liquidity,
information, and infrequently traded stocks. Journal of Finance 51,
1405–1436.
\item Grammig, J., Theissen, E., Zehnder, L.S., 2015. Bayesian Estimation of the
Probability of Informed Trading. Conference on Financial Econometrics &
Empirical Asset Pricing 2016, Lancaster University
}
}
\seealso{
\itemize{
\item \code{\link[=simulateEKOP]{simulateEKOP()}} for simulating trades data without mis-specification
\item \code{\link[=correct_trades]{correct_trades()}} to reconstruct original trades data from mis-specified
ones
}
}
