\name{bayespin-package}
\alias{bayespin-package}
\alias{bayespin}
\docType{package}
\title{
\packageTitle{bayespin}
}
\description{
\packageDescription{bayespin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayespin}
\packageIndices{bayespin}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bayespin}

Maintainer: \packageMaintainer{bayespin}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
